Update IPermutation
byte --> int

+++ DotNetTransformer/Math/Group/Permutation/IPermutation.cs
@@ -2,9 +2,9 @@ using System.Collections.Generic;
 
 namespace DotNetTransformer.Math.Group.Permutation {
 	public interface IPermutation<T> : IFiniteGroupElement<T>
-		, IEnumerable<byte>
+		, IEnumerable<int>
 		where T : IPermutation<T>
 	{
-		byte this[int index] { get; }
+		int this[int index] { get; }
 	}
 }
